Change a magic number to somehow make it work
When `sid` is set to 3 is works for the norwegians but lth gets info
from almost exactly 3 years ago. Changing the option to seemingly any
other number makes it work (but seems to break it for the norwegians,
hence the new argument).

Maybe ask the norwegians how they found the json api url, I have no
clue.